feat: add rgb light effects and name the layers
The liatris microcontroller is equipped with RGB lights on the top, this
makes it possible to change the color depending on the layer, this is
just a clue and more like a trick than a feature, but it's easier than
making the light blink a certain number of times depending on the layer.

This also names the layers with an enum which makes it clearer to know
which layer is being referenced.

The cmd key is now on the left side to keep it consistent across layers,
it will return to the base on any layer and work as cmd if held.